dcs-setup: Updating dockerfile and package names.
- Renamed package from uk.gov.cabinetoffice.di to uk.gov.di.
- Only copying a single jar over for docker.
