+++ .gitignore
@@ -28,3 +28,5 @@ hs_err_pid*
 
 **/certs/
 **/keys/
+
+build/
