--- src/main/java/uk/gov/cabinetoffice/di/ipv/dcsatpservice/DcsAtpServiceApplication.java
@@ -1,13 +1,0 @@
-package uk.gov.cabinetoffice.di.ipv.dcsatpservice;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class DcsAtpServiceApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(DcsAtpServiceApplication.class, args);
-    }
-
-}
