--- src/main/java/uk/gov/cabinetoffice/di/ipv/dcsatpservice/domain/DcsCheckRequestDto.java
@@ -1,24 +1,0 @@
-package uk.gov.cabinetoffice.di.ipv.dcsatpservice.domain;
-
-import com.fasterxml.jackson.annotation.JsonFormat;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-import java.time.Instant;
-
-@Getter
-@Setter
-@NoArgsConstructor
-public class DcsCheckRequestDto {
-    
-    private String passportNumber;
-    private String surname;
-    private String[] forenames;
-
-    @JsonFormat(pattern = "yyyy-MM-dd'T'HH:mm:ss", timezone = "UTC")
-    private Instant dateOfBirth;
-
-    @JsonFormat(pattern = "yyyy-MM-dd'T'HH:mm:ss", timezone = "UTC")
-    private Instant expiryDate;
-}
