--- src/main/java/uk/gov/cabinetoffice/di/ipv/dcsatpservice/services/dcs/AtpService.java
@@ -1,9 +1,0 @@
-package uk.gov.cabinetoffice.di.ipv.dcsatpservice.services.dcs;
-
-import reactor.core.publisher.Mono;
-import uk.gov.cabinetoffice.di.ipv.dcsatpservice.domain.DcsCheckRequestDto;
-
-public interface AtpService {
-
-    Mono<String> checkPassportData(DcsCheckRequestDto dto);
-}
