--- src/main/java/uk/gov/cabinetoffice/di/ipv/dcsatpservice/services/dcs/EncryptionService.java
@@ -1,10 +1,0 @@
-package uk.gov.cabinetoffice.di.ipv.dcsatpservice.services.dcs;
-
-import reactor.core.publisher.Mono;
-
-public interface EncryptionService {
-
-    Mono<String> encrypt(String data);
-
-    Mono<String> decrypt(String data);
-}
