--- src/main/java/uk/gov/cabinetoffice/di/ipv/dcsatpservice/services/dcs/SigningService.java
@@ -1,10 +1,0 @@
-package uk.gov.cabinetoffice.di.ipv.dcsatpservice.services.dcs;
-
-import reactor.core.publisher.Mono;
-
-public interface SigningService {
-
-    Mono<String> signData(String data);
-
-    Mono<String> unwrapSignature(String data);
-}
