--- src/main/java/uk/gov/cabinetoffice/di/ipv/dcsatpservice/services/dcs/impl/DcsEncryptionServiceImpl.java
@@ -1,30 +1,0 @@
-package uk.gov.cabinetoffice.di.ipv.dcsatpservice.services.dcs.impl;
-
-import org.springframework.beans.factory.annotation.Qualifier;
-import org.springframework.stereotype.Service;
-import reactor.core.publisher.Mono;
-import uk.gov.cabinetoffice.di.ipv.dcsatpservice.services.dcs.AbstractEncryptionService;
-
-import java.security.Key;
-import java.security.cert.Certificate;
-
-@Service("dcs-encryption-service")
-public class DcsEncryptionServiceImpl extends AbstractEncryptionService {
-
-    public DcsEncryptionServiceImpl(
-        @Qualifier("dcs-client-encryption-key") Key dcsClientEncryptionKey,
-        @Qualifier("dcs-server-encryption-cert") Certificate dcsServerEncryptionCert
-    ) {
-        super(dcsClientEncryptionKey, dcsServerEncryptionCert);
-    }
-
-    @Override
-    public Mono<String> encrypt(String data) {
-        return super.encrypt(data);
-    }
-
-    @Override
-    public Mono<String> decrypt(String data) {
-        return super.decrypt(data);
-    }
-}
