--- src/main/java/uk/gov/cabinetoffice/di/ipv/dcsatpservice/services/dcs/mock/DcsMockServerEncryptionServiceImpl.java
@@ -1,30 +1,0 @@
-package uk.gov.cabinetoffice.di.ipv.dcsatpservice.services.dcs.mock;
-
-import org.springframework.beans.factory.annotation.Qualifier;
-import org.springframework.stereotype.Service;
-import reactor.core.publisher.Mono;
-import uk.gov.cabinetoffice.di.ipv.dcsatpservice.services.dcs.AbstractEncryptionService;
-
-import java.security.Key;
-import java.security.cert.Certificate;
-
-@Service("mock-server-encryption-service")
-public class DcsMockServerEncryptionServiceImpl extends AbstractEncryptionService {
-
-    public DcsMockServerEncryptionServiceImpl(
-        @Qualifier("dcs-mock-server-encryption-key") Key dcsClientEncryptionKey,
-        @Qualifier("dcs-client-encryption-cert") Certificate dcsServerEncryptionCert
-    ) {
-        super(dcsClientEncryptionKey, dcsServerEncryptionCert);
-    }
-
-    @Override
-    public Mono<String> encrypt(String data) {
-        return super.encrypt(data);
-    }
-
-    @Override
-    public Mono<String> decrypt(String data) {
-        return super.decrypt(data);
-    }
-}
