+++ src/main/java/uk/gov/di/ipv/DcsAtpServiceApplication.java
@@ -1,0 +1,13 @@
+package uk.gov.di.ipv;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class DcsAtpServiceApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(DcsAtpServiceApplication.class, args);
+    }
+
+}
