+++ src/main/java/uk/gov/di/ipv/domain/DcsCheckRequestDto.java
@@ -1,0 +1,24 @@
+package uk.gov.di.ipv.domain;
+
+import com.fasterxml.jackson.annotation.JsonFormat;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+import java.time.Instant;
+
+@Getter
+@Setter
+@NoArgsConstructor
+public class DcsCheckRequestDto {
+    
+    private String passportNumber;
+    private String surname;
+    private String[] forenames;
+
+    @JsonFormat(pattern = "yyyy-MM-dd'T'HH:mm:ss", timezone = "UTC")
+    private Instant dateOfBirth;
+
+    @JsonFormat(pattern = "yyyy-MM-dd'T'HH:mm:ss", timezone = "UTC")
+    private Instant expiryDate;
+}
