+++ src/main/java/uk/gov/di/ipv/domain/DcsCheckResponseDto.java
@@ -1,0 +1,12 @@
+package uk.gov.di.ipv.domain;
+
+import lombok.Builder;
+import lombok.Getter;
+
+@Getter
+@Builder
+public class DcsCheckResponseDto {
+    
+    private final boolean passportValid;
+    private final String[] errorMessages;
+}
