+++ src/main/java/uk/gov/di/ipv/domain/Thumbprints.java
@@ -1,0 +1,11 @@
+package uk.gov.di.ipv.domain;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+@Getter
+@AllArgsConstructor
+public class Thumbprints {
+    private String sha1Thumbprint;
+    private String sha256Thumbprint;
+}
