+++ src/main/java/uk/gov/di/ipv/services/dcs/SigningService.java
@@ -1,0 +1,10 @@
+package uk.gov.di.ipv.services.dcs;
+
+import reactor.core.publisher.Mono;
+
+public interface SigningService {
+
+    Mono<String> signData(String data);
+
+    Mono<String> unwrapSignature(String data);
+}
