+++ src/main/java/uk/gov/di/ipv/services/dcs/impl/AtpSigningServiceImpl.java
@@ -1,0 +1,34 @@
+package uk.gov.di.ipv.services.dcs.impl;
+
+import com.google.gson.Gson;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.beans.factory.annotation.Qualifier;
+import org.springframework.stereotype.Service;
+import reactor.core.publisher.Mono;
+import uk.gov.di.ipv.domain.Thumbprints;
+import uk.gov.di.ipv.services.dcs.AbstractSigningService;
+
+import java.security.Key;
+
+@Slf4j
+@Service("atp-ipv-signing-service")
+public class AtpSigningServiceImpl extends AbstractSigningService {
+
+    public AtpSigningServiceImpl(
+        @Qualifier("atp-ipv-signing-key") Key atpIpvSigningKey,
+        @Qualifier("atp-ipv-signing-thumbprints") Thumbprints atpIpvSigningThumbprints,
+        Gson gson
+    ) {
+        super(atpIpvSigningKey, null, atpIpvSigningThumbprints, gson);
+    }
+
+    @Override
+    public Mono<String> signData(String data) {
+        return super.signData(data);
+    }
+
+    @Override
+    public Mono<String> unwrapSignature(String data) {
+        return super.unwrapSignature(data);
+    }
+}
