+++ src/main/java/uk/gov/di/ipv/utils/ClasspathFileReader.java
@@ -1,0 +1,17 @@
+package uk.gov.di.ipv.utils;
+
+import java.io.*;
+
+public abstract class ClasspathFileReader {
+    public static Reader getResourceFileReader(String fileName) throws IOException {
+        ClassLoader classLoader = ClassLoader.getSystemClassLoader();
+        try (InputStream is = classLoader.getResourceAsStream(fileName)) {
+            if (is == null) return null;
+            try (InputStreamReader isr = new InputStreamReader(is);
+                 BufferedReader reader = new BufferedReader(isr)) {
+                return reader;
+//                return reader.lines().collect(Collectors.joining(System.lineSeparator()));
+            }
+        }
+    }
+}
