--- src/test/java/uk/gov/cabinetoffice/di/ipv/dcsatpservice/DcsAtpServiceApplicationTests.java
@@ -1,13 +1,0 @@
-package uk.gov.cabinetoffice.di.ipv.dcsatpservice;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class DcsAtpServiceApplicationTests {
-
-    @Test
-    void contextLoads() {
-    }
-
-}
