+++ src/test/java/uk/gov/di/ipv/DcsAtpServiceApplicationTests.java
@@ -1,0 +1,13 @@
+package uk.gov.di.ipv;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class DcsAtpServiceApplicationTests {
+
+    @Test
+    void contextLoads() {
+    }
+
+}
